requirements.txt
fastapi
uvicorn
twilio
python-dotenv
websockets
httpx
groq
